+++ freechat-service/src/main/java/fun/freechat/service/organization/impl/OrgServiceImpl.java
@@ -1,10 +1,15 @@
 package fun.freechat.service.organization.impl;
 
+import static org.mybatis.dynamic.sql.SqlBuilder.isEqualTo;
+import static org.mybatis.dynamic.sql.SqlBuilder.isIn;
+
 import fun.freechat.mapper.OrgRelationshipDynamicSqlSupport;
 import fun.freechat.mapper.OrgRelationshipMapper;
 import fun.freechat.model.OrgRelationship;
 import fun.freechat.service.organization.OrgService;
 import fun.freechat.util.graph.Graph;
+import java.util.*;
+import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.ibatis.session.SqlSession;
@@ -13,29 +18,25 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import java.util.*;
-import java.util.stream.Collectors;
-
-import static org.mybatis.dynamic.sql.SqlBuilder.isEqualTo;
-import static org.mybatis.dynamic.sql.SqlBuilder.isIn;
-
 @Service
 @Slf4j
 @SuppressWarnings("unused")
 public class OrgServiceImpl implements OrgService {
     @Autowired
     private SqlSessionFactory sqlSessionFactory;
+
     @Autowired
     OrgRelationshipRepo orgRelationshipRepo;
+
     @Autowired
     OrgRelationshipMapper orgRelationshipMapper;
 
     private boolean isEnabled(OrgRelationship relationship) {
-        return relationship.getEnabled().equals((byte)1);
+        return relationship.getEnabled().equals((byte) 1);
     }
 
     private boolean isVirtual(OrgRelationship relationship) {
-        return relationship.getIsVirtual().equals((byte)1);
+        return relationship.getIsVirtual().equals((byte) 1);
     }
 
     @Override
@@ -51,26 +52,26 @@ public class OrgServiceImpl implements OrgService {
             int rows;
             for (String ownerId : new HashSet<>(owners)) {
                 boolean isVirtual = solidSubordinates.contains(ownerId);
-                OrgRelationship oldRelationship = orgRelationshipMapper.selectOne(c ->
-                        c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(userId))
+                OrgRelationship oldRelationship = orgRelationshipMapper
+                        .selectOne(c -> c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(userId))
                                 .and(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(ownerId)))
                         .orElse(null);
                 if (oldRelationship != null) {
                     if (isEnabled(oldRelationship) && isVirtual == isVirtual(oldRelationship)) {
                         continue;
                     }
-                    rows = orgRelationshipMapper.updateByPrimaryKeySelective(
-                            oldRelationship.withGmtModified(now)
-                                    .withEnabled((byte)1)
-                                    .withIsVirtual(isVirtual ? (byte)1 : (byte)0));
+                    rows = orgRelationshipMapper.updateByPrimaryKeySelective(oldRelationship
+                            .withGmtModified(now)
+                            .withEnabled((byte) 1)
+                            .withIsVirtual(isVirtual ? (byte) 1 : (byte) 0));
                 } else {
                     OrgRelationship relationship = new OrgRelationship()
                             .withGmtCreate(now)
                             .withGmtModified(now)
                             .withUserId(userId)
                             .withOwnerId(ownerId)
-                            .withEnabled((byte)1)
-                            .withIsVirtual(isVirtual ? (byte)1 : (byte)0);
+                            .withEnabled((byte) 1)
+                            .withIsVirtual(isVirtual ? (byte) 1 : (byte) 0);
                     rows = orgRelationshipMapper.insert(relationship);
                 }
                 if (rows > 0) {
@@ -104,26 +105,26 @@ public class OrgServiceImpl implements OrgService {
             int rows;
             for (String subordinateId : new HashSet<>(subordinates)) {
                 boolean isVirtual = solidOwners.contains(subordinateId);
-                OrgRelationship oldRelationship = orgRelationshipMapper.selectOne(c ->
-                                c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(subordinateId))
-                                        .and(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(userId)))
+                OrgRelationship oldRelationship = orgRelationshipMapper
+                        .selectOne(c -> c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(subordinateId))
+                                .and(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(userId)))
                         .orElse(null);
                 if (oldRelationship != null) {
                     if (isEnabled(oldRelationship) && isVirtual == isVirtual(oldRelationship)) {
                         continue;
                     }
-                    rows = orgRelationshipMapper.updateByPrimaryKeySelective(
-                            oldRelationship.withGmtModified(now)
-                                    .withEnabled((byte)1)
-                                    .withIsVirtual(isVirtual ? (byte)1 : (byte)0));
+                    rows = orgRelationshipMapper.updateByPrimaryKeySelective(oldRelationship
+                            .withGmtModified(now)
+                            .withEnabled((byte) 1)
+                            .withIsVirtual(isVirtual ? (byte) 1 : (byte) 0));
                 } else {
                     OrgRelationship relationship = new OrgRelationship()
                             .withGmtCreate(now)
                             .withGmtModified(now)
                             .withUserId(subordinateId)
                             .withOwnerId(userId)
-                            .withEnabled((byte)1)
-                            .withIsVirtual(isVirtual ? (byte)1 : (byte)0);
+                            .withEnabled((byte) 1)
+                            .withIsVirtual(isVirtual ? (byte) 1 : (byte) 0);
                     rows = orgRelationshipMapper.insert(relationship);
                 }
                 if (rows > 0) {
@@ -149,9 +150,8 @@ public class OrgServiceImpl implements OrgService {
         if (StringUtils.isBlank(userId) || CollectionUtils.isEmpty(owners)) {
             return false;
         }
-        int rows = orgRelationshipMapper.delete(c ->
-                c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(userId))
-                        .and(OrgRelationshipDynamicSqlSupport.ownerId, isIn(owners)));
+        int rows = orgRelationshipMapper.delete(c -> c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(userId))
+                .and(OrgRelationshipDynamicSqlSupport.ownerId, isIn(owners)));
         return rows > 0;
     }
 
@@ -160,8 +160,8 @@ public class OrgServiceImpl implements OrgService {
         if (StringUtils.isBlank(userId) || CollectionUtils.isEmpty(subordinates)) {
             return false;
         }
-        int rows = orgRelationshipMapper.delete(c ->
-                c.where(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(userId))
+        int rows =
+                orgRelationshipMapper.delete(c -> c.where(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(userId))
                         .and(OrgRelationshipDynamicSqlSupport.userId, isIn(subordinates)));
         return rows > 0;
     }
@@ -171,13 +171,13 @@ public class OrgServiceImpl implements OrgService {
         if (StringUtils.isBlank(userId)) {
             return false;
         }
-        TreeSet<String> subordinates = (TreeSet<String>) orgRelationshipRepo.getSubordinates(userId, false)
-                .tailSet(userId);
+        TreeSet<String> subordinates = (TreeSet<String>)
+                orgRelationshipRepo.getSubordinates(userId, false).tailSet(userId);
         boolean successful = false;
         for (String subordinateId : subordinates.descendingSet()) {
-            int rows = orgRelationshipMapper.delete(c ->
-                    c.where(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(subordinateId)));
-            successful =  rows > 0 || successful;
+            int rows = orgRelationshipMapper.delete(
+                    c -> c.where(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(subordinateId)));
+            successful = rows > 0 || successful;
         }
         return successful;
     }
@@ -190,11 +190,12 @@ public class OrgServiceImpl implements OrgService {
         if (owners == null) {
             owners = Collections.emptyList();
         }
-        Set<String> banOwners = orgRelationshipMapper.select(c ->
-                        c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(userId)))
-                .stream()
-                .map(OrgRelationship::getOwnerId)
-                .collect(Collectors.toSet());
+        Set<String> banOwners =
+                orgRelationshipMapper
+                        .select(c -> c.where(OrgRelationshipDynamicSqlSupport.userId, isEqualTo(userId)))
+                        .stream()
+                        .map(OrgRelationship::getOwnerId)
+                        .collect(Collectors.toSet());
         banOwners.removeAll(owners);
         boolean successful = removeOwners(userId, banOwners);
         successful = (!CollectionUtils.isEmpty(addOwners(userId, owners))) || successful;
@@ -209,11 +210,12 @@ public class OrgServiceImpl implements OrgService {
         if (subordinates == null) {
             subordinates = Collections.emptyList();
         }
-        Set<String> banSubordinates = orgRelationshipMapper.select(c ->
-                        c.where(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(userId)))
-                .stream()
-                .map(OrgRelationship::getUserId)
-                .collect(Collectors.toSet());
+        Set<String> banSubordinates =
+                orgRelationshipMapper
+                        .select(c -> c.where(OrgRelationshipDynamicSqlSupport.ownerId, isEqualTo(userId)))
+                        .stream()
+                        .map(OrgRelationship::getUserId)
+                        .collect(Collectors.toSet());
         banSubordinates.removeAll(subordinates);
         boolean successful = removeSubordinates(userId, banSubordinates);
         successful = (!CollectionUtils.isEmpty(addSubordinates(userId, subordinates))) || successful;
@@ -225,8 +227,8 @@ public class OrgServiceImpl implements OrgService {
         if (StringUtils.isBlank(userId) || StringUtils.isBlank(otherUserId)) {
             return false;
         }
-        return getOwners(userId, true).contains(otherUserId) ||
-                getSubordinates(userId, true).contains(otherUserId);
+        return getOwners(userId, true).contains(otherUserId)
+                || getSubordinates(userId, true).contains(otherUserId);
     }
 
     @Override
